# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubernetes REST API",
    "version": "1.0.0",
    "description": "A simplified REST API for managing Kubernetes resources using client-go. Supports pods, deployments, services, namespaces, configmaps, persistent volume claims, nodes, and cluster info."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/deployments": {
      "get": {
        "summary": "List deployments",
        "description": "Returns all deployments, optionally filtered by namespace.",
        "operationId": "listDeployments",
        "tags": ["Deployments"],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Filter by namespace.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of deployment summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DeploymentSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create a deployment",
        "description": "Creates a Deployment from a simplified request body.",
        "operationId": "createDeployment",
        "tags": ["Deployments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDeploymentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeploymentSummary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/deployments/{namespace}/{name}": {
      "get": {
        "summary": "Inspect a deployment",
        "description": "Returns the full Kubernetes Deployment object.",
        "operationId": "getDeployment",
        "tags": ["Deployments"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Deployment object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeploymentSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a deployment",
        "description": "Removes a deployment and its managed pods.",
        "operationId": "deleteDeployment",
        "tags": ["Deployments"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "204": { "description": "Deployment deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/deployments/{namespace}/{name}/scale": {
      "post": {
        "summary": "Scale a deployment",
        "description": "Sets the desired replica count for the deployment.",
        "operationId": "scaleDeployment",
        "tags": ["Deployments"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScaleRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated scale",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScaleResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/deployments/{namespace}/{name}/restart": {
      "post": {
        "summary": "Rolling restart a deployment",
        "description": "Triggers a rolling restart by patching the pod template annotation with the current timestamp.",
        "operationId": "restartDeployment",
        "tags": ["Deployments"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Restart initiated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/configmaps": {
      "get": {
        "summary": "List configmaps",
        "operationId": "listConfigMaps",
        "tags": ["ConfigMaps"],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of configmap summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ConfigMapSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create a configmap",
        "operationId": "createConfigMap",
        "tags": ["ConfigMaps"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateConfigMapRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ConfigMap created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ConfigMapSummary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/configmaps/{namespace}/{name}": {
      "get": {
        "summary": "Inspect a configmap",
        "description": "Returns the full ConfigMap including all data keys and values.",
        "operationId": "getConfigMap",
        "tags": ["ConfigMaps"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "ConfigMap object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ConfigMapSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a configmap",
        "operationId": "deleteConfigMap",
        "tags": ["ConfigMaps"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "204": { "description": "ConfigMap deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "namespace": {
        "name": "namespace",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Kubernetes namespace"
      },
      "name": {
        "name": "name",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource name"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or missing required fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Kubernetes API error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "pod not found" }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "rolling restart initiated" }
        }
      },
...
      "DeploymentSummary": {
        "type": "object",
        "properties": {
          "name":               { "type": "string" },
          "namespace":          { "type": "string" },
          "replicas":           { "type": "integer" },
          "ready_replicas":     { "type": "integer" },
          "available_replicas": { "type": "integer" },
          "updated_replicas":   { "type": "integer" },
          "image":              { "type": "string" },
          "labels":             { "type": "object", "additionalProperties": { "type": "string" } },
          "created_at":         { "type": "string", "format": "date-time" },
          "conditions":         { "type": "array", "items": { "type": "string" } }
        }
      },
      "EnvVar": {
        "type": "object",
        "properties": {
          "name":  { "type": "string" },
          "value": { "type": "string" }
        },
        "required": ["name", "value"]
      },
      "CreateDeploymentRequest": {
        "type": "object",
        "required": ["name", "image"],
        "properties": {
          "name":      { "type": "string", "example": "nginx" },
          "namespace": { "type": "string", "default": "default" },
          "replicas":  { "type": "integer", "default": 1 },
          "image":     { "type": "string", "example": "nginx:latest" },
          "port":      { "type": "integer", "example": 80 },
          "labels":    { "type": "object", "additionalProperties": { "type": "string" } },
          "env": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EnvVar" }
          }
        }
      },
      "ScaleRequest": {
        "type": "object",
        "required": ["replicas"],
        "properties": {
          "replicas": { "type": "integer", "minimum": 0, "example": 3 }
        }
      },
      "ScaleResponse": {
        "type": "object",
        "properties": {
          "name":      { "type": "string" },
          "namespace": { "type": "string" },
          "replicas":  { "type": "integer" }
        }
      },
...
      "ConfigMapSummary": {
        "type": "object",
        "properties": {
          "name":       { "type": "string" },
          "namespace":  { "type": "string" },
          "keys":       { "type": "array", "items": { "type": "string" } },
          "labels":     { "type": "object", "additionalProperties": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateConfigMapRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name":      { "type": "string" },
          "namespace": { "type": "string", "default": "default" },
          "data":      { "type": "object", "additionalProperties": { "type": "string" } },
          "labels":    { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
...
    }
  },
...
}